Touchscreen pointercal work
With InputDevice.java looking for pointercal in /data/system/tslib,
and ts_calibrate and friends looking for pointercal in that same
location, we want to install our product specific pointercal to
/system/etc/pointercal and have ins_pointercal.sh (ran from
init.q1u.rc) to copy it to /data/system/tslib.

Work will be done later to the Home apk or init to check for a
touchscreen device, and if it finds one, it should check for
/system/etc/pointercal. If /system/etc/pointercal or /data/system/tslib
do not exist, then it should automatically run ts_calibrate.